+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Jira.js changelog
 
+### 2.0.2
+
+- `Buffer.from` replaced to raw JS code in Basic authorization
+- Telemetry config type fixed
+- `noCheckAtlassianToken` flag added to config (`X-Atlassian-Token: no-check`)
+- Typing improves
+
 ### 2.0.1
 
 - Types bug fixes